Refactor code structure and update imports in chess module
