+++ backend/chess/chess.py
@@ -2,13 +2,13 @@ from typing import Union
 
 import contants.colors as Colors
 import contants.info_keys as InfoKeys
-import chess.moves as Moves
-import chess.pieces as Pieces
-import chess.rewards as Rewards
+import contants.rewards as Rewards
+import contants.moves as Moves
+import pieces as Pieces
 import gym
 import numpy as np
 import pygame
-from chess.types import Cell
+from models.types import Cell, Action, Trajectory
 from gym import spaces
 from pygame.font import Font
 from pygame.surface import Surface
