+++ backend/chess/contants/__init__.py
@@ -1,2 +1,4 @@
 from .colors import *
 from .info_keys import *
+from .moves import *
+from .rewards import *
