+++ backend/chess/contants/rewards.py
