+++ backend/chess/models/__init__.py
@@ -1,1 +1,1 @@
-
+from .types import *
